feat(front): separate concerns with a new component for playlist dropdown features
