+++ src/front/js/component/favouritePlaylistsRender.jsx
@@ -1,40 +1,23 @@
-import React, { useContext, useState } from "react";
-import {
-  Accordion,
-  Button,
-  Col,
-  Container,
-  Dropdown,
-  ListGroup,
-  ListGroupItem,
-  Row,
-  ThemeProvider
-} from "react-bootstrap";
-import DropdownToggle from "react-bootstrap/esm/DropdownToggle";
+import React from "react";
+import { Accordion, Col, Container, ListGroup, ListGroupItem, Row, ThemeProvider } from "react-bootstrap";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faPlay, faEllipsisVertical } from "@fortawesome/free-solid-svg-icons";
+import { faPlay } from "@fortawesome/free-solid-svg-icons";
 
-import { Context } from "../store/appContext";
+import { FavouritePlaylistDropdownFeats } from "./favouritePlaylistsDropdownFeats.jsx";
 
 import { msToMin } from "./utils/msToMin";
 
 
 export const FavouritePlayistsRender = ({listOfPlaylists}) => {
-    const {store, actions} = useContext(Context);
-    const [dropdownState, setDropdownState] = useState(false);
 
-    const handleClickMoveTrack = (tracks, index, playlistOrigin, playlistDestiny) => {
-        actions.moveToPlaylist("playlistStore", tracks, index, playlistOrigin, playlistDestiny)
-    };
-
-    const playlistComponentBuilderII = (
+    const playlistAccordion = (
         <Accordion>
             {listOfPlaylists?.map((playlist) =>                 
                 <Accordion.Item key={playlist.id} eventKey={playlist.id}>
                     <Accordion.Header className="d-flex">
                         <div className="btn btn-success me-3" variant="success"><FontAwesomeIcon icon={faPlay} /></div>
-                            <strong>{playlist.name}</strong> &nbsp;
-                            <em>- {playlist.tracks.length} songs</em>  
+                        <strong>{playlist.name}</strong> &nbsp;
+                        <span>- {playlist.tracks.length} songs</span>  
                     </Accordion.Header>
                           
                     <Accordion.Body>
@@ -68,34 +51,7 @@ export const FavouritePlayistsRender = ({listOfPlaylists}) => {
                                     </Col>  
                                     
                                     <Col xs="2" sm="1" md="1" lg="1">
-                                        <Dropdown autoClose={dropdownState}>
-
-                                            <DropdownToggle className="ms-1 me-auto" variant="outline-success">
-                                                <FontAwesomeIcon icon={faEllipsisVertical} />
-                                            </DropdownToggle>
-                                            <Dropdown.Menu>
-
-                                                <Dropdown.Item onClick={() => (
-                                                    actions.addToFavourites("favouritesStore", tracks)
-                                                    )}> Add to Favourites
-                                                </Dropdown.Item>
-
-                                                <Dropdown.Item onClick={() => (
-                                                    actions.removeFromPlaylist("playlistStore", index, playlist.name) 
-                                                    )}> Remove from this playlist                                               
-                                                </Dropdown.Item>                                                
-
-                                                {listOfPlaylists.map((playlistName, index) => 
-                                                    {if (playlistName.name !== playlist.name) {
-                                                        return (
-                                                            <Dropdown.Item key={index} onClick={() => handleClickMoveTrack(tracks, index, playlist.name, playlistName.name)}>
-                                                                Move to "{playlistName.name}"                                                       
-                                                            </Dropdown.Item>
-                                                        )}                                                
-                                                    }
-                                                )}
-                                            </Dropdown.Menu>
-                                        </Dropdown>
+                                        <FavouritePlaylistDropdownFeats playlist={playlist} listOfPlaylists={listOfPlaylists} />
                                     </Col>
                                      
 
@@ -117,7 +73,7 @@ export const FavouritePlayistsRender = ({listOfPlaylists}) => {
                     <Col
                         md={{ span: 10, offset: 1 }}
                         lg={{ span: 10, offset: 1 }}>
-                        {playlistComponentBuilderII}
+                        {playlistAccordion}
                     </Col>
                 </Row>                
             </ThemeProvider>
