+++ src/front/js/store/flux.js
@@ -21,10 +21,6 @@ const getState = ({ getStore, getActions, setStore }) => {
 
     actions: {
       // Use getActions to call a function within a fuction
-      exampleFunction: () => {
-        getActions().changeColor(0, "green");
-      },
-
       getMessage: async () => {
         try {
           // fetching data from the backend
